Code corrections were made.

+++ src/main/java/com/Project/accountManager/dto/request/AccountDepositRequest.java
@@ -2,7 +2,9 @@ package com.Project.accountManager.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AccountDepositRequest {
-    int depositAmount;
+    BigDecimal depositAmount;
 }
